+++ get_icon.py
@@ -1,0 +1,19 @@
+from .get_background import download_image_with_httpx_auto_redirect
+import asyncio
+
+# BACKGROUND_URL = "https://www.loliapi.com/acg/"
+# DEFAULT_ICON = "./minecraft-creeper-face.png"
+
+async def get_icon_image(url: str):
+    if url.startswith("http"):
+        icon_data = await download_image_with_httpx_auto_redirect(url)
+        if icon_data:
+            return icon_data
+        else:
+            return None
+    else:
+        def read_file(path):
+            with open(path, "rb") as f:
+                return f.read()
+        loop = asyncio.get_event_loop()
+        return await loop.run_in_executor(None, read_file, url)
